+++ src/battle/battle_trainer.gd
@@ -6,6 +6,8 @@ var name: String = "Trainer" ## The trainer's name
 var team: PokemonTeam ## The trainer's team
 var is_player: bool = false ## True if the trainer is the player.
 var is_wild: bool = false ## True if it's a wild pokemon.
+var back_frames: SpriteFrames ## SpriteFrames to show if the trainer is an ally.
+var front_sprite: Texture2D ## Sprite to show if the trainer is an enemy. Unused if [member is_wild] is true.
 
 
 func _init(trainer_name: String, trainer_team: PokemonTeam, trainer_is_player: bool = false, trainer_is_wild: bool = false) -> void: